chore(si-sdf): de-noise output & harden ws stream consuming (#215)
This change is mostly about de-noising the logging output of `si-sdf`,
namely demoting a lot of tracing info to....`trace` levels :)

However, a second part of this change is updating the web socket
streaming logic to handle all inner variants of a `warp::ws::Message`.
While running the system, I found that there were more messages being
consumed than the `Message::text` types of messages that we are
expecting. In my case, typically extra `Message::close` messages. This
lead me to read the warp and tungstenite code a bit and refactor the
consuming code to handle all possible message types.

Finally, the precondition state of testing a billing account signup is
fixed, allowing a developer to run the
`filters::billing_accounts::create` test more than once in a row.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>