chore: make deploy//internal (#216)
Fixes [ch637].

This commit adds a revised docker configuration for all the components,
released to docker hub. Each container can be built with:

```
make container//components/si-web-app
```

Where `si-web-app` could be any docker-able component. It also adds a
Pulumi configuration to automatically deploy a full EKS cluster, along
with the resources, needed to get a working internal deployment up at

```
https://internal.systeminit.com
```

This deployment, along with the production deployment, are all currently
password protected.

To make a release to the internal deployment, you do:

```
make deploy//internal
```

It will build the containers, tag and push them, and then run a pulumi
up with the release as an argument. The results will be a newly running
k8s cluster.