chore: changing properties recalculates successors
Fixes [ch793].

When we change a property on an entity, we expect that any successor
entity who would get an inferred property from it should recalculate.
That's what this PR does.

It fixes it for both initial creation (so that the node you see
instantly has the right values on create) and on subsequent edits.

It's 100% unoptimized, and probably has some edge cases where things can
still go wrong (I suspect if you rely on successor recalculation before
an action, where the action is in the middle of the change set, things
won't be right when you run the action?).

Lets fix those bugs when we come to them, though. Today, let us
celebrate how glorious it is to see things refresh automatically.