feat(sdf): add encrypted secrets to API (#220)
This change introduces the concept of secrets to the system. Currently,
secrets are not too dissimilar from Entities in that a `Secret` has an
object type and kind which are opaque to the backend. That is, the
message component of the secret is arbitrary data which the front end
will interpret and use as structured data.

A `Secret` currently has one object type of `Credential` with room to
create other types in the future (perhaps single value secret strings,
for example). Additionally, a `Secret` has a kind which helps the API
consumer to understand the message format. Currently there is one kind
which is `DockerHub`, likely consisting of a `username` and `password`
hash.

Secrets are encrypted client side (in our case, the front end in the
browser) and send to the API on creation using the active billing
account's current public key (which is also new to the system). The API
only ever exposes a `Secret` type which boils down to its name and basic
type metadata, such as `object_type` and `kind`. Notably, neither the
encypted payload nor the unencrypted message will be exposed by the API,
but rather used internally when preparing to invoke JavaScript functions
in `veritech`. For this, an `EncryptedSecret` can be fetched from the
database and then decrypted via a fallible type conversion in Rust into
a `DecryptedSecret`.

The secrets API allows a secret to be created, fetched individually, and
as a list. Notably, the API currently does not support deleting a
`Secret` nor updating one and are therefore assumed to be immutable
(until we decide otherwise).

The public key API is a child under the billing accounts path, i.e.
`billingAccounts/{id}/publicKey`. This call fetches the single current
public key that should be used to encrypt any secrets. A simple
`rotate_public_key` method exists on `BillingAccount` to support future
key rotation so our front end code should be ready to use a new key
whenever encrypting a new secret.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>