feat: add kubernetes secret object (#221)
Closes [ch855].

This PR adds a kubernetes secret object. It defaults to the Opaque
secret type, and supports both stringData and the base64 required data
field.

It does not implement the docker hub secret integration, but that makes
sense (we aren't there yet).

It refactors the shared kubernetes code into a single kubernetesShared
library.