feat: kubernetesDeployment gets imagePullSecrets from kubernetesSecrets (#227)
Fixes ch[854].

This PR finishes the kubernetesSecret node for creating docker pull
credentials. If you wire the kubernetesSecret to a dockerHubCredential,
then we automatically populate the kubernetesSecret with the appropriate
credential. We're also careful not to leak the secret.

When you then link the kubernetesSecret to a kubernetesDeployment, it
automatically adds each appropriate kubernetesSecret to the
imagePullSecrets array.