feat: connect to AWS EKS clusters correctly (#229)
Fixes [ch864], [ch854], and [ch868].

You can now add an AWS credential, an AWS EKS cluster, and wire them up
to kubernetes objects. When you do, we will then automatically configure
the entities and sync the state!

It's amazing.

It also refactors the SecretList vue component into a List and a New
component, with the New component loading specific sub-components for
the given secret data.