fix(si-registry): fix `dockerImage`'s sync when image is not pulled (#230)
It looks like somewhere in refactoring we omitted the step of `docker
image pull` in the `syncResource` intelligence. In that case we aren't
fully exercising the `dockerHubCredential`s nor getting the benefit of
updated metadata if the tag in the repository has updated (i.e. tracking
`latest`).

Note that this fix in the implementation might look a touch more
verbose, but I'm hoping that a Go-style check if error and return style
might help us catch this the next time and be rebase friendly when the
eventing/logging work lands.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>