fix(si-web-app): handle undefined `resource` in `SquareChart`
There is a loading flow (which I haven't 100% nailed down) where a
Service's `resource` is not found in the editor state. Our computed
`resource()` returns `Resource | undefined` so that is a legit state and
we were not handling this in the template. One more class of front end
console errors gone.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>