fix(si-sdf): when calculating successor props, fall back to head (#233)
This fixes an issue when executing a changeset (real or hypothetical)
where a successor is *not* a participant in the changeset. Prior to this
change, the code to calculate the properties of successors called
`Entity::get_projection` which would fail with a `NoHead` error and thus
causing the execution to abort. This change recognizes that a successor
might not be found in the change set, and if that is the case, it uses
the `head` version of Entity.

References: [ch975]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>