feat(si-web): EventBar shows events and EventLogs
This updates the EventBar to show events and EventLogs. It doesn't
support showing the outputlines yet - rather than spend time hooking
that up, we should wait until we refactor the full event code (we should
make shared components, rather than duplicate all the code we know we're
going to refactor).

This fixes [ch936].