feat: add awsIamJsonPolicy to model IAM JSON policies (#245)
This implementation takes a simple/naive approach by modeling the JSON
document as...a JSON document. In other words, there is one primary
property called `object` which is rendered as an HTML `<textarea>` and
contains the JSON payload directly. A related property of `objectJson`
is a `PropCode` property and will be rendered in the `CodeViewer`
component.

References: [ch937]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>