feat: Add a CLI server and client, using websockets (#246)
Fixes [ch1014].

This creates a server and minimal client for taking commands, executing
them, and returning messages over a websocket. The CLI submits a
Command, which is then executed by SDF. The resulting events, event
logs and output lines are then streamed back to the CLI, and passed
through a formatter.

It's very minimal, but it should work!