feat(sdf, web, cli): add API clients
This PR fully adds working API clients. Fixes [ch972].

You can log in to the Web UI, click the "Clients" menu, and add a new
CLI client. Once it is added, it shows you your token in a buffer. That
token will never be shown again, so make sure you copy it. :)

You can then use that token to authenticate with the CLI.

This also splits authentication in two - you can now authenticate as a
User to the Web UI, or as an API Client to via the CLI. You cannot
re-use the tokens between the two. (by design, for now).