feat(si): trigger a node action from a new CLI, `si` (#254)
This change adds a new CLI program `si` with a `node run` subcommand
which will trigger an action on a given Node in a System and optionally
set multiple properties on arbitrary Nodes. Our canonical example is to
trigger a `deploy` action on a service Node and set the `image` property
on a dockerImage. The Nodes are identified using their nodeId.

References: [ch973]

Building
--------

To compile a development build, you can run the following from the root
of the project or in the `components/si-sdf` subdirectory:

```console
$ cargo build --bin si
```

Likely the CI release build will be:

```console
$ cargo build --bin si --release --verbose
$ cp ./target/release/si /path/to/dst/
```

Example
-------

For example, assuming an `SI_TOKEN` environment variable is set with a suitable
client token,

```console
$ si node run \
  --set node:ff6031e63bdd48b88090de73276d0b83.image=nginx:latest \
  system:d872097c2ef54fa1ad102fd5b514329e \
  node:14a3861d171f4268ab8d6abc0040ba88 \
  deploy
```

Top level short help
--------------------

```console
$ si -h
si 0.1.0
The System Initiative <info@systeminit.com>

The command line client for the System Initiative

USAGE:
    si <SUBCOMMAND>

OPTIONS:
    -h, --help       Prints help information
    -V, --version    Prints version information

SUBCOMMANDS:
    help    Prints this message or the help of the given subcommand(s)
    node    Manage nodes
```

Top level long help
-------------------

```console
$ si --help
si 0.1.0
The System Initiative <info@systeminit.com>

The command line client for the System Initiative

https://www.systeminit.com/

USAGE:
    si <SUBCOMMAND>

OPTIONS:
    -h, --help
            Prints help information

    -V, --version
            Prints version information

SUBCOMMANDS:
    help
            Prints this message or the help of the given subcommand(s)

    node
            Manage nodes
```

Node run subcommand short help
------------------------------

```console
$ si node run -h
si-node-run
Runs an action on a node in a system

USAGE:
    si node run <SYSTEM_ID> <NODE_ID> <ACTION> --token <token>

ARGS:
    <SYSTEM_ID>    System id for the node
    <NODE_ID>      Target node id
    <ACTION>       Action to invoke [ex: deploy]

OPTIONS:
    -F, --formatter <formatter>
            Output formatter to use [env: SI_OUTPUT_FORMATTER] [default:
            simple] [possible values: debug, simple]

    -h, --help                     Prints help information
    -H, --host <host>
            SI Client API host [env: SI_HOST] [default: ws://localhost:5156]

    -s, --set <set>...
            Set a property on a node [ex:
            node:bd345f3ec2dd41d3867d02bd9d5e2d8d.image=nginx:1.19.5]

    -T, --token <token>            SI Client token [env: SI_TOKEN]
    -V, --version                  Prints version information
```

Node run subcommand short help
------------------------------

```console
$ si node run --help
    Finished dev [unoptimized + debuginfo] target(s) in 0.23s
     Running `target/debug/si node run --help`
si-node-run
Runs an action on a node in a system

Optionally, a set of properties can be set on specific nodes which allows you
to update versions, tags, etc. for a deployment for example.

USAGE:
    si node run <SYSTEM_ID> <NODE_ID> <ACTION> --token <token>

ARGS:
    <SYSTEM_ID>
            System id for the node
    <NODE_ID>
            Target node id
    <ACTION>
            Action to invoke [ex: deploy]

OPTIONS:
    -F, --formatter <formatter>
            Output formatter to use [env: SI_OUTPUT_FORMATTER] [default:
            simple] [possible values: debug, simple]

    -h, --help
            Prints help information

    -H, --host <host>
            SI Client API host [env: SI_HOST] [default: ws://localhost:5156]

    -s, --set <set>...
            Set a property on a node [ex:
            node:bd345f3ec2dd41d3867d02bd9d5e2d8d.image=nginx:1.19.5]

    -T, --token <token>
            SI Client token [env: SI_TOKEN]

    -V, --version
            Prints version information
```

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>