bug(si-sdf): Resources and Events need new txns (#268)
This PR fixes bugs caused by Events and Resources not existing (yet),
while resource syncronization needs to update both items. To solve it,
we threaded the pg pool and nats through to events and resources,
allowing the system to write them outside the transaction created by the
handler.

This commit leaves in a significant amount of error debugging, since
we're pretty sure we're not out of the proverbial woods yet.

It includes a test that triggers the issue, but whose module is
commented out. This is because we will need to update the design to
allow for optional callbacks when these async calls have completed, so
the reactor can block rather than being torn down immediately.

Love,
Adam, Fletcher, and Alex.