bug: resources can be created before nodes
Resources exist outside a handlers transaction model, which means that
they could be created before the node/entity they reference exists. This
change makes it so we don't throw an error if that happens - instead, we
just return.

It's probably overbroad, in that it just swallows all errors - but short
of refactoring for a different model of error handling altogether, I
think it's good enough for government work.