bug: unknown property 'language' breaks code editor
The code editor would break if an object didn't have a codeProperty
value set on it. Since not every object does, if the first object that
gets loaded doesn't have one, it would break the CodeViewer component,
and the UI would not recover.

This fix just makes it so we take the codeProperty being null as the
language also being null, and viola - working code editor.