bug: do not send blank queries to the backend
A slight change from n1ql to pgsql is that an 'AND' with '()' as its
argument is rejected by the lexer, rather than accepted as an empty
clause. This adds a check to ensure if the user sends us a query that
evaluates to nothing, we don't try and add nothing as a condition on
listing the model.