bug: close sending task when send channel is gone (#276)
We detect when the remote side of a websocket disappears, but we don't
abort the sending task when we get a send error. This PR makes it so we
print a debug line (which, I suspect, we won't need for long - just
until we're sure this is the right behavior), and breaks out of the
sending thread.

The result is that you see 1 error message when a websocket is closed
and the sender still exists, but none after that.