bug: billing account signup should create default system
The billing account signup process didn't create a default system.
Instead, it waited for the frontend to do it - which meant that both new
testing in the frontend *and* behavior in the backend were strangely
mismatched.

This PR fixes that. It also means that a changeSet and editSession get
created and committed in billing account new - so if those paths break
now, well... everything does.

Good times.

Love,
Adam