bug: fix the code editor, add component testing (#281)
This PR switches the code editor from Monaco back to Codemirror, and
adds an initial pass at a component-level testing framework using
Javascripts Testing Library for Vue, along with a bunch of magical
sausage to get things to work.

It adds a CodeMirror.vue test, which does a fine job of testing that the
component renders the code properties correctly. Due to the nature of
codemirror, it's hard to test the write path - but I did test it
manually.

The editor also is (almost) the entire vertical width (if it's the full
width and you have enough YAML, things get very strange), has a
relatively matching theme, goes read only when not in edit mode (or if
you don't have a code property on the node you're viewing), and no
longer displays any of the strange loading/unloading/popping in and out
behavior.