bug: remove 401 error on node creation
We were seeing a 401 Unauthorized error on node creation. This was from
the associated Resource object being created (and pushed down the
websocket) before the Node itself was committed. That resulted in the
Update code dispatching the saved Resource to the application store,
which then tries to update the list of resources on the application
list. When it can't find the node, it caught the exception, but chrome
automatically logs the 401.

This commit refactors the node get handler to return a 404 if the
resource doesn't exist, moving the tenancy check to *after* the fetch.
This way we know we're getting the right error, at least.

It also updates the code in the application store to use a new
`get_cache` method. This doesn't do the fallback get, and instead only
returns if it has been cached. The side effects should be minimal, since
the next time we see the resource the node will have been cached. The
result is there is no more error in the console, and if there is, we
will get the right error code back regardless.