chore: refactor signup and login
This PR sets up a refactored architecture, using the Signup
and Login components as our examples. It moves the old web ui into
`si-web-app-old`.

The new design flips the relationship between the frontend and the
backend - it assumes the frontend's perspective of the API it wants is
correct, and the backend provides specific endpoints for the behavior
needed by the frontend.

The front end now consists of:

Pages (the full frame of the page)
 |-Views (a segment of a page; think nav)
   |-Wads (a 'wad' of components on the page)

It also now has:

Api
 |-SDF
   |- model/* (pure data/some function models of backend objects)
   |- dal/* (data access layers)

The job the DAL objects is to tie together the requests needed by
the frontend and the data model on the backend. So a common flow would
be:

Page
 |-View
   |-Wad
     |-Store
       |-DAL
         |-Model

If there is nothing to persist for an action, a Wad may call the DAL
directly.

The side effects of this design are pretty great. The first is that it
makes testing the application in layers easy - you can see the tests for
the Login and Signup Wad's in the tree - they run fast, and are easy to
write. Each DAL becomes a specific API, which the store can use to
manage the applications state. Everything is done with dependency
injection, so we can mock the various pieces when we want to. That
should enable us to, over time, let folks completely mock a backend
interface using types that are equivalent, but without needing to
actually dispatch to the backend. (As of now, the tests are not mocked)

This PR also stats a new pedant test, the "demo" test. This is the
primary test we will add on to as we make progress against the demo, and
serves as the end to end test proving the product does what we expect.

Love,
Adam and Fletcher