chore: refactor panels
This commit refactors the Web UI to use a new structure for Panels. It
starts with a PanelTree, which takes a datastructure that consists of
PanelContainers and Panels. Each PanelContainer has an orientation,
either "row" or "column", and can contain either panels or another
panelContainer. Panels can be added when the mouse is over a container
with 'alt+shift+n', and the last panel can be deleted with
'alt+shift+d'. You cannot delete the last panel in a panelcontainer, and
there is a maximum of 2 panels per container.

Panels are sized with percentages, so that they hold their shape and
size when the browser window is itself resized. They can be resized
within their container, and can fill a maximum of 70% of the available
space and a minimum of 30%.

This PR also adds in two pieces of infrastructure needed to implement
these (and I expect many future) features: the PartyBus and the
PanelEventBus.

First up is PartyBus - this is used to emit events between various store
instances without causing too much deep coupling. It will be used by the
store instances themselves to receive and emit events, and by the Update
Websocket to signal changes to the stores. When a user selects a new
changeSet, for example, it is the PartyBus that sends the event to all
subscribers.

The PanelEventBus is a Vue component event bus. It is used to send
events to multiple components within the PanelTree (and perhaps expands
to be across the UI, which will require a rename.) This allows us to
have the UI Components react to changes without needing a strict
parent/child relationship, or to deal with too much event forwarding.
When a user asks for a new panel via a keyboard shortcut, it is the
PanelEventBus that dispatches the shortcut event and the panel-create
event.  Use it with care and love.