chore: draft attribute panel, session store persist
This commit has a draft attribute panel, and adds a central Persister
that can be used to persist state, either to the URL or to session
storage. There are still some state flow bugs, but in general, it's much
better than it was before.

If you need to wipe out the session store, you can either close your
tab, or do "sessionStore.clear()" in the javascript debug console.