chore: refactor attribute and code panels
This brings the attribute panel and a working code view into the system.
It's most done - the outstanding issues are refresh events between
changed code views, and the proper deletion of repeated objects and map
entries when changed from the code view.

It also has a lot of my comments left in it - we'll clean all that up
later. :)