chore(sdf): add node_position model
This is a first step addition of a NodePosition model to track the x/y
positions of Nodes within a "contex"--that is, a uniquely named
identifier that allows one Node to have multiple positions in different
contexts.

Currently, the `positions` field on `Node` is removed, which breaks the
current contract with the front end, but that will be addressed in
future work.

Additionally, several unused warnings were cleaned up.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>