build: enable clean run of root `make build`
This change update some of the Make targets to allow a reasonably
successful run of `make build` from the root of the project. I say
reasonably today because we presently have several TypeScript production
compilation errors which prevent `si-web-app` from building to
completion. As a result of this, I pushed the building of the web app
until the end, after building SDF as our backend should likely *always*
be buildable.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>