feat: entity qualification (#315)
This PR implements the first pass at entity qualifications. It allows
you to specify a list of qualifications in the schema for an entity, and
then provide a callback which will execute to validate that entity.

We run the qualifications whenever an entity is updated, and we run them
if a predecessor is updated (although there is currently a bug where we
are not threading through the system correctly, which will eventually
bite us - but is probably fine for this second.)

The UI updates in real time, shows status errors. Works goodish.

Love,
Adam and Fletcher