feat: workflows and resources
This commit adds in the concept of 'workflows', and specifically wires
together the 'actions' an entity can be asked to perform (which map to
workflows). Additionally, it brings 'resources' back in to the model,
but this time as strictly a run-time concern (you no longer have a
resource that gets updated in an edit session).

The workflow code itself is a little dense, but has some nice patterns.
In particular the 'invoke' tasks being able to be either stand-alone
async tasks or awaited via a one-shot channel is pretty great.

Workflows themselves are expressed as pure data in the registry,
alongside the schema for entities.