feat: attribute editing functional pass (#333)
This work is a functional pass of the Attribute editor. It works for
complex objects - you should be able to nest any kind of supported data
structure/widget combination in any way you want, and it should edit it
correctly. It has minimal styling - there is quite a bit that should be
done to make it both more approachable and more ergonomic. But it
fucking works, and that's pretty amazing.

It breaks the various components of the AttributeViewer up into many
smaller components, which then are recursively loaded. The side effect
is that it should be easy to add new kinds of widgets, or to customize
them with different behavior - without having to think too much about
the internals of how our Entities work.

This pass also fixes some bugs in the override logic, so that things now
are truly selectable via tombstoning (and don't impact anything other
than the layer you have tombstoned.) Tombstones should work regardless
of the structure of the data, although visualizing it might not.

Things that aren't implemented but should be when we have the rest of
the end-to-end together:

* Goldbars
* Tweaking the edit counts to reflect only visible changes
* A ton of how things are laid out when nested
* In partcular, how to deal with nested object labels
* Some bugs where we're causing the browser to overflow completely.

But the core editing works!

Love,
Adam