feat: deployment and component schematics (#336)
This PR adds working versions of the deployment and component
schematics. It includes migrating the Schematic code from Vuex to RxJS,
updating and extending the logic for menus, and making the backend be
the authoritative source for the various types of schematics that can be
returned.

It also has a few nice-to-haves, like turning on logging for every query
in postgres.

Excelsior!