feat(*): conceptual nodes and implementations
This PR adds the idea of a conceptual node with an implementation. It
allows you to specify that a node is conceptual, and specify an
implementation property for it. That property then makes any available
implementation edge (meaning you have connected a valid implementation
node) to be selected in the attribute editor for that node.

When you select an implementation, it is shown directly on the node
itself on the deployment schematic.

It also adds a couple more stub schemas for Kubernetes and AWS EKS, so
that you can link two together. It exposes the fact that you probably
need to futher validate edges between nodes in a deployment diagram, to
ensure that a given implementation is accepatble to a given deployment
target. I'll leave that open for now.

Excelsior!