feat(*): you can deploy the demo application end to end
This PR makes it so you can deploy the demo application end to end. It
requires a few things to be installed on your machine:

* kubeval
* kubectl
* aws
* docker

But you can now create a high level service node, its implementation,
wire it up to deploy on a kubernetes cluster, and then deploy the whole
thing onto a cloud provider instance.

Took 6:30 if we have to add the secrets in. 5:31 if you don't!

Long pole is AWS DNS provisioning.

It is amazing.

Also, SI was probably really born on my wife's birthday, which is fun!