fix(si-web-app): session state is persisted
This adds a `SessionBroker` component, which wraps pretty much the
entire application. It watches observables, writes the values to the
browsers sessionStorage, and restores them when it is first created. The
side effect is that you don't really have to worry all that much anymore
about state tracking across refreshes - if you want something to
refresh, you can add it to the SessionBroker, and let it handle the rest
for you.

There is some tricky business in when you should clear state - but it's
not pervasive. Essentially, the highest "owning" component should ensure
that it clears out the observables that are no longer needed.

In addition to watching the observables like `changeSet` or
`editSession`, this PR also watches the panel state for you. It will
remember the panels selected for a given application for the duration of
the session.

In order to make this all work, the final days of the Vuex store have
also arrived. The store, the panelBus, the updateTracker, and the
persister have all been condemned to the dustbin of history. We salute
you, thousands of lines of code, for you served valiantly and well. But
it was time to go.

Best,
Adam