feat: implement "Deploy" button services summary
This change adds behavior to the "Deploy" button in the services summary
section in the application context. It's job it to run the "deploy"
action on the current application. The "application:deploy" workflow
finds all directly included "service" entities and calls their
"service:deploy" action.

Currently our selector strategies are rather limited--that is all
selected nodes are run in serial--but once we have parallel/concurrent
execution...woah!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>