feat(si-web): add menu is organized
This PR fixes [ch1174]. It organizes the Node Add Menu according to an
arbitrarily nested category, and is filtered both by the kind of
schematic and the root object of that schematic.

New nodes fill in their menu entry in their schema, and can appear in
multiple schematics, in multiple places, with arbitrary names. Pretty
much we can classify anything we want into whatever menu we want.

It also improves the node add menu:

* It now has an artifically wide "target" for mouse over, making it
  easier to directly access the menu without it closing.
* There is a half-second delay between the mouse leaving the menu and
  the menu automatically closing.
* If the mouse re-enters the menu within that half second delay, the
  menu will not close.

May the new menus live long and prosper.