feat(si-web): implement save on enter
This PR fixes [ch1173].

When in the attribute editor, all the "text" fields (text, numbers,
etc.) will now unfocus and save when you hit 'enter'.

When you are in a select box, we now automatically save the field when
you make a selection. We do not unfocus the select box, though.

When you are in a text area (currently only accessible via the torture
test node) you can save and unfocus the current field with either
"shift+enter" or "alt+enter".