feat(si-web): green field names for driven fields (#360)
This PR partially solves [ch1186].

It turns (most) fields that are inferred green. It doesn't show you
where it comes from - but it does show that the value is inferred from
another source.

Currently if you are in a deeply nested array, things go off the rails,
for reasons I'm not enitrely sure of, but think has to do with how we
collapse the empty positions in existing arrays.