feat(si-web): deploy to azure
This PR kicks the extra point of being able to deploy to either AWS or
Azure. There is no intelligence what-so-ever in the Azure resources yet,
but if you set it all up correctly, you can deploy to AKS.

Additionally, it very much hacks support for deploying to multiple
clusters at once. Right now, it does that by just looking to see if
there are multiple clusters present in the context for the command on
Veritech, and if there are, running the command multiple times. This
might even be the right approach, depending on what we think we should
do with resources. (Are they always 1:1?)

But it works!