feat(web, sdf): basic support for revisions
This PR adds basic support for revisions, and it closes [ch1129].

The revisions menu gets populated with a chronological list of previous
changeSets. Selecting one will show the state of entities as they
existed at the time of the changeSet being applied. It disables
selection of the changeSet menu, and makes sure that no other buttons
allowing editing are accessible.

What it doesn't do is show you the full state of the system *at the time
the change set was applied* - instead it overlays the changeSet on head.
That's because we don't take a full snapshot, and we probably should, so
that we can show you how the world was then.

But it's good enough for now.