feat(web, sdf): resource sync and implementation (#369)
This PR implements resources for all of our entities, and periodic sync.
It fixes [ch1219].

The specific behavior for a given entity lives in the si-veritech intel
functions in a `syncResource` callback. It is responsible for both
populating runtime data about the resource and checking its health.

The scheduler is very simple - it tries to schedule all the resources as
closet to every 30 seconds as possible. It will run long, but it will
never be more often than 30 seconds.

Some resources also have sub-resources. If these are present, then the
key in the sub-resource hash is the entity id of a cluster they are
attached to.

This PR requires wiping your database, so be ready for that. :)

Best,
Adam