fix(web): node teleportation
This PR fixes [ch1171]. The cause of this was that we used the component
state to record the condition we wanted the world to be in, and then we
would dispatch async functions that expect that world not to change when
they hit an await point. That assumption (an easy one to make) is wrong
- in fact, the data can, and does, sometimes change. The result was the
kind of "teleportation" we saw.

Unfortunately, it causes there to be an occasional "jitter" now when you
are moving a node around. I think I know the approximate cause (that the
calls are not resolved in order), but it's not a trivial fix. This
removes the big problem (things flying off the screen) for a smaller one
(nodes occasionally "jumping" for a split second sometimes).