feat: terminate services & delete Kubernetes objects
This change adds a `termintate` action and workflow on `service` nodes
that work through their implementation nodes for Kubernetes services. In
this case, the `delete` action is called on Kubernetes objects
(triggering a `kubectl delete` command to run) in a prescribed order
similar to that of Helm for uninstalling (similar to our "terminate").

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>