feat(web): Kubernetes image pull policy (#387)
This PR makes the imagePullPolicy for a k8sDeployment be inferred from
the shape of the docker image string. If it does not have a tag, or the tag is
'latest', then the imagePullPolicy is set to 'Always'. If there is a tag
set to anything else, it is set to 'IfNotPresent'.

It is also now a required field.